+++ tests/model.test.cjs
@@ -21,3 +21,4 @@ Module._extensions['.ets'] = (module, filename) => module._compile(ts.transpileM
   compilerOptions: { target: ts.ScriptTarget.ES2021, module: ts.ModuleKind.CommonJS }
 }).outputText, filename);
 require('../fold_motion/src/test/FoldMotionModel.test.ets').default();
+require('../fold_motion/src/test/FoldMotionDirectionModel.test.ets').default();